Refactor(#48): QuizService 클래스 분리
